backend/pyproject.toml
[tool.poetry]
name = "voting-system-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "project"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.0"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"